Improved process scheduler

+++ include/sys/stddef.h
@@ -1,6 +1,6 @@
 #ifndef	VIOS_STDDEF_H
 #define	VIOS_STDDEF_H
 
-#define offsetof(type, member) ((size_t) &((type *)0)->member)
+#define offsetof(type, member) ((u32) &((type *)0)->member)
 
 #endif
